Add player cube prefab to test camera controller.
Add playermovement to control test cube movement.

Create camera controller(Only work in orthographic mode).

--UIChunter
