+++ Zoridia/Assets/Materials/Player1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4a7af1d3ccc28f49992dba36d821d1d
+timeCreated: 1465341843
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
