+++ Zoridia/Assets/Prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: ff99d4152e86b73468a89d6e045f6e86
+folderAsset: yes
+timeCreated: 1465339975
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
