+++ Zoridia/Assets/Prefab/Player1.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb244a23a47c07743b94275b8b37a91a
+timeCreated: 1465339978
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
