+++ Zoridia/Assets/Scripts/Camera.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2b3a993d0a9c97446b39cf234fe9b541
+folderAsset: yes
+timeCreated: 1465344005
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
