+++ Zoridia/Assets/Scripts/Camera/CameraController.cs
@@ -1,0 +1,106 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class CameraController : MonoBehaviour {
+	public float m_DampTime = 0.2f;                 
+    public float m_ScreenEdgeBuffer = 4f;           
+    public float m_MinSize = 6.5f;                  
+    /*[HideInInspector]*/ public Transform[] m_Targets; //TODO: GameManager transfer player gameobject to m_Targets. 
+
+
+    private Camera m_Camera;                        
+    private float m_ZoomSpeed;                      
+    private Vector3 m_MoveVelocity;                 
+    private Vector3 m_DesiredPosition;              
+
+
+    private void Awake()
+    {
+        m_Camera = GetComponentInChildren<Camera>();
+    }
+
+
+    private void FixedUpdate()
+    {
+        Move();
+        Zoom();
+    }
+
+
+    private void Move()
+    {
+        FindAveragePosition();
+
+        transform.position = Vector3.SmoothDamp(transform.position, m_DesiredPosition, ref m_MoveVelocity, m_DampTime);
+    }
+
+	//Find the middle point position of two targets.
+    private void FindAveragePosition()
+    {
+        Vector3 averagePos = new Vector3();
+        int numTargets = 0;
+
+        for (int i = 0; i < m_Targets.Length; i++)
+        {
+            if (!m_Targets[i].gameObject.activeSelf)
+                continue;
+
+            averagePos += m_Targets[i].position;
+            numTargets++;
+        }
+
+        if (numTargets > 0)
+            averagePos /= numTargets;
+
+        //In this version freeze camera position vertically.
+        //If our players can fly, we can unlock the vertical freezing.
+        averagePos.y = transform.position.y;
+
+        m_DesiredPosition = averagePos;
+    }
+
+
+    private void Zoom()
+    {
+        float requiredSize = FindRequiredSize();
+		m_Camera.orthographicSize = Mathf.SmoothDamp(m_Camera.orthographicSize, requiredSize, ref m_ZoomSpeed, m_DampTime);
+    }
+
+
+    private float FindRequiredSize()
+    {
+        Vector3 desiredLocalPos = transform.InverseTransformPoint(m_DesiredPosition);
+
+        float size = 0f;
+
+        for (int i = 0; i < m_Targets.Length; i++)
+        {
+            if (!m_Targets[i].gameObject.activeSelf)
+                continue;
+
+            Vector3 targetLocalPos = transform.InverseTransformPoint(m_Targets[i].position);
+
+            Vector3 desiredPosToTarget = targetLocalPos - desiredLocalPos;
+
+            size = Mathf.Max (size, Mathf.Abs (desiredPosToTarget.y));
+
+            size = Mathf.Max (size, Mathf.Abs (desiredPosToTarget.x) / m_Camera.aspect);
+        }
+        
+        size += m_ScreenEdgeBuffer;
+
+        size = Mathf.Max(size, m_MinSize);
+
+        return size;
+    }
+
+
+    public void SetStartPositionAndSize()
+    {
+        FindAveragePosition();
+
+        transform.position = m_DesiredPosition;
+
+        m_Camera.orthographicSize = FindRequiredSize();
+    }
+}
