+++ Zoridia/Assets/Scripts/Camera/CameraController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 05413dcdbf6a43e40b848b93f437ed89
+timeCreated: 1465344020
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
