+++ Zoridia/Assets/Scripts/Player.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b8f319fa7fc365a40a4ceef1a3073f60
+folderAsset: yes
+timeCreated: 1465343969
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
