+++ Zoridia/Assets/Scripts/Player/PlayerMovement.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 03ae36910aaf5014a8ba35cdc16392f2
+timeCreated: 1465340137
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
